press 'F2' to reset options to default settings
- does not work for slider items
- does work for list items, but doesn't show it
- also I left in some commented code for a "reset to default button",
but took it out because of the errors I mentioned. If anyone runs into
an issue where the screen height is too big, tell them to use F2 in the
options screen.
- also, the "height bigger than actual screen" bug should only happen if
you have at least 2 monitors attached, and are playing on one that has
less height than the other